requirements.txt
openfeature-sdk==0.4.1
rox==5.0.10
